Solidity Smart Contract
Smart contract that deposits, withdraws, and transfers eth.  Any transaction which is performed in the Bank.sol contract, relays data back to an external contract named 'Financial.sol via the financial.sol 'get transaction' public view returns function.  

Ownable.sol is inherited into the bank.sol contract via 'import' which holds the initiated and onlyOwner values of the contract.

'Destroy_contract.sol' is the kill switch to terminante the contract and in theory send any ether in the contract to msg.sender wallet.